Fix semver; better error if no spaceID given
When backups are created (or when helper.GetSpaceFromContext is called)
there was no error that the space ID was not found.
Two errors were created:
no spacefile and missing team ID
This might be confusing to the user, since providing a spaceID fixes the
problem as well.
Also, this change switches to a different semver implementation, since
the first one lacked support for tilde.