Add online validation
SpaceCTL will use the API to check if the application ID and version is
available.
Can be disabled with -o/--offline. To make this work, all Validate()
functions now require the `offline bool` parameter. However, only the
last method call utilises this parameter.
This commit also introduces the application show method which lists all
available versions for this application.